split completion tests over default and custom test classes for better readability
