+++ test/completion/custom/completion.js
